Enemy Projectiles Complete (For Now)
I have finished the enemy's shooting, allowing them to fire at the enemy from set points with a set range & angle, dealing damage when the bullet hits.

Note: The rotation of the fire angle wasn't playing nice, so I removed it for the time being.

Next Up: Overhaul of the Weapon Room.