Bugfixes, QOL & Footsteps
Fixed a few bugs.
Changed the layout of the ship to better fit the stations.
The player now rotates to face the direction they are walking in
The player now has footsteps
Upped the damage of weapon 3 & added a 4th weapon at the back of the ship
Moved the shield to better protect the new ship layout